+++ NEWS.md
@@ -40,6 +40,7 @@ roundness <- function(matrix) {
  * The correct types of changes are now plotted in legend when plotting the results of `test.metric`.
  * `get.disparity` now correctly concatenates one dimensional results into a `"numeric"` (rather than a `"matrix"`).
  * `make.metric` now internally handles `covar` object correctly (i.e. as distance matrices).
+ * Calculating disparity for multiple matrices and customised subsets now works as expected for all metric levels.
 
 dispRity v1.7 (2022-08-08) *MacMacGlimm*
 =========================